+++ notification/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
     preset: 'ts-jest',
     testEnvironment: 'node',
-    roots: ['./tests']
+    roots: ['./src']
 }
